Integrate download.servo.org into the website
This patch reuses code from download.servo.org to show the date of the
last builds.
It removes the current table at download.servo.org,
but it seems kind of duplicated with the content above it.
It includes the instructions that were present at download.servo.org
for macOS and Linux under servo.org/download