Stop using scores-last-run.json
The only point of that file was if it meant that scores.json could use a
more efficient format (including less detail for older runs).

That is not the case due to the tooltip in the chart needing the full
data for older runs, so the separate json file is pure overhead.

Signed-off-by: Nico Burns <nico@nicoburns.com>